refactor(builtins): changed linked list to dynamic array

+++ src/dyn_array/dyn_array.h
@@ -5,8 +5,8 @@
 #include <stdint.h>
 
 typedef struct {
-  uint64_t cap;
-  uint64_t len;
+  size_t cap;
+  size_t len;
   void **arr;
 } DynamicArray;
 
@@ -17,9 +17,7 @@ typedef struct {
 } IntArray;
 
 void int_array_append(IntArray *, int);
-
 void int_array_init(IntArray *, size_t);
-
 void int_array_free(IntArray *);
 
 void array_append(DynamicArray *, void *);
